metasploit-mcp-server/requirements.txt
mcp
requests
python-dotenv
pymetasploit3
fastapi
uvicorn